Reorganized the snapshot.py file into the snapshot module. (#67)
Beyond breaking up the snapshot.py file into separate files, the following changes were also made:
  * Created the write_json_file within serializers/utils.py. This was to remove all json module things from the snapshot module. Updated tests for this change.
  * Added a test for trying to snapshot when not within a context. This was a TODO that never got done and took the opportunity.
  * Created module-level imports so that none of these changes are backwards incompatible.
