+++ snappiershot/snapshot/__init__.py
@@ -1,0 +1,4 @@
+""" Module relating to snapshots and the public API. """
+from .metadata import SnapshotMetadata
+from .snapshot import Snapshot
+from .status import SnapshotStatus
