+++ tests/test_serializers/test_io.py
@@ -2,7 +2,11 @@
 import json
 
 import pytest
-from snappiershot.serializers.io import SnapshotKeys, parse_snapshot_file
+from snappiershot.serializers.io import (
+    SnapshotKeys,
+    parse_snapshot_file,
+    write_json_file,
+)
 
 
 class TestParseSnapshotFile:
@@ -52,3 +56,24 @@ class TestParseSnapshotFile:
         # Act & Assert
         with pytest.raises(ValueError):
             parse_snapshot_file(snapshot_file)
+
+
+class TestWritingToFile:
+    """ Tests for the file writing utility functions. """
+
+    @staticmethod
+    def test_write_json_error(tmp_path):
+        """ Test if an error occurs during snapshot writing, no file is written. """
+        # Arrange
+        obj = {("tuple as key",): None}
+        snapshot_file = tmp_path / "snapshot_file.json"
+
+        # Act
+        try:
+            write_json_file(obj, snapshot_file)
+        except TypeError:
+            pass
+
+        # Assert
+        assert not snapshot_file.exists()
+        assert not list(snapshot_file.parent.glob("*"))
